+++ lib/asset/computeplan_validation.go
@@ -10,6 +10,8 @@ import (
 func (t *NewComputePlan) Validate() error {
 	return validation.ValidateStruct(t,
 		validation.Field(&t.Key, validation.Required, is.UUID),
+		validation.Field(&t.Tag, validation.Length(0, 100)),
+		validation.Field(&t.Name, validation.Required, validation.Length(1, 100)),
 		validation.Field(&t.Metadata, validation.By(validateMetadata)),
 	)
 }