+++ server/standalone/dbal/computeplan.go
@@ -20,6 +20,7 @@ type sqlComputePlan struct {
 	DeleteIntermediaryModels bool
 	CreationDate             time.Time
 	Tag                      string
+	Name                     string
 	Metadata                 map[string]string
 	TaskCounts               persistence.ComputePlanTaskCount
 }
@@ -32,6 +33,7 @@ func (cp *sqlComputePlan) toRawComputePlan() *asset.ComputePlan {
 		DeleteIntermediaryModels: cp.DeleteIntermediaryModels,
 		CreationDate:             timestamppb.New(cp.CreationDate),
 		Tag:                      cp.Tag,
+		Name:                     cp.Name,
 		Metadata:                 cp.Metadata,
 	}
 }
@@ -72,8 +74,8 @@ func (d *DBAL) ComputePlanExists(key string) (bool, error) {
 func (d *DBAL) AddComputePlan(plan *asset.ComputePlan) error {
 	stmt := getStatementBuilder().
 		Insert("compute_plans").
-		Columns("key", "channel", "owner", "delete_intermediary_models", "creation_date", "tag", "metadata").
-		Values(plan.Key, d.channel, plan.Owner, plan.DeleteIntermediaryModels, plan.CreationDate.AsTime(), plan.Tag, plan.Metadata)
+		Columns("key", "channel", "owner", "delete_intermediary_models", "creation_date", "tag", "name", "metadata").
+		Values(plan.Key, d.channel, plan.Owner, plan.DeleteIntermediaryModels, plan.CreationDate.AsTime(), plan.Tag, plan.Name, plan.Metadata)
 
 	return d.exec(stmt)
 }
@@ -81,7 +83,7 @@ func (d *DBAL) AddComputePlan(plan *asset.ComputePlan) error {
 // GetComputePlan returns a ComputePlan by its key
 func (d *DBAL) GetComputePlan(key string) (*asset.ComputePlan, error) {
 	stmt := getStatementBuilder().
-		Select("key", "owner", "delete_intermediary_models", "creation_date", "tag", "metadata", "task_count", "waiting_count", "todo_count", "doing_count", "canceled_count", "failed_count", "done_count").
+		Select("key", "owner", "delete_intermediary_models", "creation_date", "tag", "name", "metadata", "task_count", "waiting_count", "todo_count", "doing_count", "canceled_count", "failed_count", "done_count").
 		From("expanded_compute_plans").
 		Where(sq.Eq{"key": key, "channel": d.channel})
 
@@ -92,7 +94,7 @@ func (d *DBAL) GetComputePlan(key string) (*asset.ComputePlan, error) {
 
 	pl := new(sqlComputePlan)
 
-	err = row.Scan(&pl.Key, &pl.Owner, &pl.DeleteIntermediaryModels, &pl.CreationDate, &pl.Tag, &pl.Metadata, &pl.TaskCounts.Total, &pl.TaskCounts.Waiting, &pl.TaskCounts.Todo, &pl.TaskCounts.Doing, &pl.TaskCounts.Canceled, &pl.TaskCounts.Failed, &pl.TaskCounts.Done)
+	err = row.Scan(&pl.Key, &pl.Owner, &pl.DeleteIntermediaryModels, &pl.CreationDate, &pl.Tag, &pl.Name, &pl.Metadata, &pl.TaskCounts.Total, &pl.TaskCounts.Waiting, &pl.TaskCounts.Todo, &pl.TaskCounts.Doing, &pl.TaskCounts.Canceled, &pl.TaskCounts.Failed, &pl.TaskCounts.Done)
 	if err != nil {
 		if errors.Is(err, pgx.ErrNoRows) {
 			return nil, orcerrors.NewNotFound("computeplan", key)
@@ -106,7 +108,7 @@ func (d *DBAL) GetComputePlan(key string) (*asset.ComputePlan, error) {
 // GetRawComputePlan returns a compute plan without its computed properties.
 func (d *DBAL) GetRawComputePlan(key string) (*asset.ComputePlan, error) {
 	stmt := getStatementBuilder().
-		Select("key", "owner", "delete_intermediary_models", "creation_date", "tag", "metadata").
+		Select("key", "owner", "delete_intermediary_models", "creation_date", "tag", "name", "metadata").
 		From("compute_plans").
 		Where(sq.Eq{"key": key, "channel": d.channel})
 
@@ -116,7 +118,7 @@ func (d *DBAL) GetRawComputePlan(key string) (*asset.ComputePlan, error) {
 	}
 
 	pl := new(sqlComputePlan)
-	err = row.Scan(&pl.Key, &pl.Owner, &pl.DeleteIntermediaryModels, &pl.CreationDate, &pl.Tag, &pl.Metadata)
+	err = row.Scan(&pl.Key, &pl.Owner, &pl.DeleteIntermediaryModels, &pl.CreationDate, &pl.Tag, &pl.Name, &pl.Metadata)
 	if err != nil {
 		if errors.Is(err, pgx.ErrNoRows) {
 			return nil, orcerrors.NewNotFound("computeplan", key)
@@ -134,7 +136,7 @@ func (d *DBAL) QueryComputePlans(p *common.Pagination, filter *asset.PlanQueryFi
 	}
 
 	stmt := getStatementBuilder().
-		Select("key", "owner", "delete_intermediary_models", "creation_date", "tag", "metadata", "task_count", "waiting_count", "todo_count", "doing_count", "canceled_count", "failed_count", "done_count").
+		Select("key", "owner", "delete_intermediary_models", "creation_date", "tag", "name", "metadata", "task_count", "waiting_count", "todo_count", "doing_count", "canceled_count", "failed_count", "done_count").
 		From("expanded_compute_plans").
 		Where(sq.Eq{"channel": d.channel}).
 		OrderBy("creation_date ASC, key ASC").
@@ -158,7 +160,7 @@ func (d *DBAL) QueryComputePlans(p *common.Pagination, filter *asset.PlanQueryFi
 	for rows.Next() {
 		pl := new(sqlComputePlan)
 
-		err = rows.Scan(&pl.Key, &pl.Owner, &pl.DeleteIntermediaryModels, &pl.CreationDate, &pl.Tag, &pl.Metadata, &pl.TaskCounts.Total, &pl.TaskCounts.Waiting, &pl.TaskCounts.Todo, &pl.TaskCounts.Doing, &pl.TaskCounts.Canceled, &pl.TaskCounts.Failed, &pl.TaskCounts.Done)
+		err = rows.Scan(&pl.Key, &pl.Owner, &pl.DeleteIntermediaryModels, &pl.CreationDate, &pl.Tag, &pl.Name, &pl.Metadata, &pl.TaskCounts.Total, &pl.TaskCounts.Waiting, &pl.TaskCounts.Todo, &pl.TaskCounts.Doing, &pl.TaskCounts.Canceled, &pl.TaskCounts.Failed, &pl.TaskCounts.Done)
 		if err != nil {
 			return nil, "", err
 		}